scentsation_ml/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
pyyaml>=6.0
joblib>=1.1.0
